64-bit arithmetic & more primitives
Having 64-bit instead of 32-bit arithmetic was particularly tricky
because, while 64-bit numbers fit perfectly in a NUM node, the binary
numeric operations needed an intermediate FN1 node which could not carry
them, because 64 bits (for two ports) + 32 bits (for type + kind) + 64
bits (for number) is larger than 128 bits. I initially had the idea of
having 28 bit ports, which allowed the FN1 node to be represented as 8
bits (type + prim func) + 28 bits (port a) + 28 bits (port b) = 128
bits. That was awful for performance and required changing the entire
style of the code. I went back to the drawing bord and noticed I could,
simply, make FN1 point to the NUM node (instead of erasing it). That
makes some rules a little bit awkward (DUP - FN1 for one), but works
amazingly smoothly. That means FN1 nodes are stored in two different
places, but they are, logically, a single node with two ports.

The code was also adapted for having more primitives. Adding more
operations is now trivial.